Externalizes all views, projection and default settings to separate files
