+++ src/defs/defaults.js
@@ -1,0 +1,6 @@
+// Default values of the ‹h3-worldmap› public properties
+export const PROPS_DEFAULTS = {
+  WORLD_GEOMETRY_SRC: "https://cdn.jsdelivr.net/npm/world-atlas@2/land-110m.json",
+  WORLD_GEOMETRY_COLL: "land",
+  PROJECTION: "orthographic"
+}
